+++ .storybook/preview.tsx
@@ -1,0 +1,32 @@
+import type { Preview } from "@storybook/react";
+import React from "react";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { initialize, mswLoader, mswDecorator } from "msw-storybook-addon";
+import "@/styles/globals.css";
+
+initialize();
+
+const queryClient = new QueryClient();
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: "^on[A-Z].*" },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  loaders: [mswLoader],
+  decorators: [
+    mswDecorator,
+    (Story) => (
+      <QueryClientProvider client={queryClient}>
+        <Story />
+      </QueryClientProvider>
+    ),
+  ],
+};
+
+export default preview;
