+++ package.json
@@ -64,6 +64,7 @@
     "eslint-plugin-tailwindcss": "^3.13.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "msw-storybook-addon": "^2.0.0-beta.1",
     "postcss": "^8.4.31",
     "prettier": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.6",
